requirements.txt
openai==2.8.1
azure-search-documents==11.6.0
python-dotenv==1.2.1
Flask==3.1.2
pyodbc==5.3.0